requirements.txt
# Web app
flask>=2.0
flask-sqlalchemy>=2.5
flask-mailman>=0.3.0

# Chatbot (intent model)
torch>=1.9
nltk>=3.6

# Video / hand exercise pipeline
torchvision>=0.10
pillow>=8.0
opencv-python>=4.5
mediapipe>=0.10
numpy>=1.19
